refactor(ftq): make ftq a separate package
