+++ src/main/scala/xiangshan/Bundle.scala
@@ -30,8 +30,10 @@ import xiangshan.frontend._
 import xiangshan.mem.{LqPtr, SqPtr}
 import xiangshan.backend.Bundles.{DynInst, UopIdx}
 import xiangshan.backend.fu.vector.Bundles.VType
-import xiangshan.frontend.{AllAheadFoldedHistoryOldestBits, AllFoldedHistories, BPUCtrl, CGHPtr, FtqPtr, FtqToCtrlIO}
-import xiangshan.frontend.{Ftq_Redirect_SRAMEntry, HasBPUParameter, IfuToBackendIO, PreDecodeInfo, RasPtr}
+import xiangshan.frontend.{AllAheadFoldedHistoryOldestBits, AllFoldedHistories, BPUCtrl, CGHPtr}
+import xiangshan.frontend.ftq.{FtqPtr, FtqToCtrlIO}
+import xiangshan.frontend.{HasBPUParameter, IfuToBackendIO, PreDecodeInfo}
+import xiangshan.frontend.ftq.Ftq_Redirect_SRAMEntry
 import xiangshan.cache.HasDCacheParameters
 import utility._
 
@@ -41,7 +43,7 @@ import chisel3.util.experimental.decode.EspressoMinimizer
 import xiangshan.backend.CtrlToFtqIO
 import xiangshan.backend.fu.NewCSR.{Mcontrol6, Tdata1Bundle, Tdata2Bundle}
 import xiangshan.backend.fu.PMPEntry
-import xiangshan.frontend.Ftq_Redirect_SRAMEntry
+import xiangshan.frontend.ftq.Ftq_Redirect_SRAMEntry
 import xiangshan.frontend.AllFoldedHistories
 import xiangshan.frontend.AllAheadFoldedHistoryOldestBits
 import xiangshan.backend.rob.RobBundles.RobCommitEntryBundle
