+++ src/main/scala/xiangshan/backend/Bundles.scala
@@ -24,6 +24,7 @@ import xiangshan.mem.{LqPtr, SqPtr}
 import yunsuan.vector.VIFuParam
 import xiangshan.backend.trace._
 import utility._
+import xiangshan.frontend.ftq.FtqPtr
 
 object Bundles {
   /**
