+++ src/main/scala/xiangshan/frontend/FrontendBundle.scala
@@ -18,6 +18,10 @@ package xiangshan.frontend
 
 import chisel3._
 import chisel3.util._
+import ftq.BpuFlushInfo
+import ftq.Ftq_Redirect_SRAMEntry
+import ftq.Ftq_RF_Components
+import ftq.FtqPtr
 import org.chipsalliance.cde.config.Parameters
 import utility._
 import utils.NamedUInt
