+++ src/main/scala/xiangshan/mem/lsqueue/LoadQueueReplay.scala
@@ -22,7 +22,7 @@ import utils._
 import utility._
 import xiangshan._
 import xiangshan.ExceptionNO._
-import xiangshan.frontend.FtqPtr
+import xiangshan.frontend.ftq.FtqPtr
 import xiangshan.backend.rob.{RobLsqIO, RobPtr}
 import xiangshan.backend.fu.fpu.FPU
 import xiangshan.backend.fu.FuConfig._
